chore: Move Skeleton component to UI directory
