--- packages/webapp-libs/webapp-core/src/components/skeleton/index.ts
@@ -1,1 +1,0 @@
-export * from './skeleton.component';
