+++ packages/webapp-libs/webapp-core/src/components/ui/__tests__/skeleton.component.spec.tsx
@@ -1,9 +1,9 @@
 import { screen } from '@testing-library/react';
 
-import { Skeleton } from '../';
 import { render } from '../../../tests/utils/rendering';
+import { Skeleton } from '../skeleton';
 
-describe('Skeleton', () => {
+describe('UI/Skeleton', () => {
   it('Should render rounded skeleton', async () => {
     render(<Skeleton data-testid="testid" className="h-12 w-12 rounded-full" />);
