+++ packages/webapp-libs/webapp-core/src/components/ui/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from '../../lib/utils';
+import { cn } from '@sb/webapp-core/lib/utils';
 
 function Skeleton({ className, ...props }: React.HTMLAttributes<HTMLDivElement>) {
   return <div className={cn('bg-muted animate-pulse rounded-md', className)} {...props} />;
