+++ packages/webapp-libs/webapp-documents/src/routes/documents/document/skeleton/skeleton.component.tsx
@@ -1,4 +1,4 @@
-import { Skeleton as SkeletonComponent } from '@sb/webapp-core/components/skeleton';
+import { Skeleton as SkeletonComponent } from '@sb/webapp-core/components/ui/skeleton';
 
 export const Skeleton = () => {
   return (
