+++ packages/webapp/plop/crud/templates/components/editItem/editItem.component.hbs
@@ -2,7 +2,7 @@ import { useMutation, useQuery } from '@apollo/client';
 import { PageHeadline } from '@sb/webapp-core/components/pageHeadline';
 import { PageLayout } from '@sb/webapp-core/components/pageLayout';
 import { Separator } from '@sb/webapp-core/components/separator';
-import { Skeleton } from '@sb/webapp-core/components/skeleton';
+import { Skeleton } from '@sb/webapp-core/components/ui/skeleton';
 import { useGenerateLocalePath } from '@sb/webapp-core/hooks';
 import { trackEvent } from '@sb/webapp-core/services/analytics';
 import { useToast } from '@sb/webapp-core/toast/useToast';
