+++ packages/webapp/plop/crud/templates/components/itemDetails/itemDetails.component.hbs
@@ -3,7 +3,7 @@ import { gql } from '@sb/webapp-api-client/graphql';
 import { PageHeadline } from '@sb/webapp-core/components/pageHeadline';
 import { PageLayout } from '@sb/webapp-core/components/pageLayout';
 import { Separator } from '@sb/webapp-core/components/separator';
-import { Skeleton } from '@sb/webapp-core/components/skeleton';
+import { Skeleton } from '@sb/webapp-core/components/ui/skeleton';
 import { useParams } from 'react-router';
 
 export const {{ pascalCase name }}DetailsQuery = gql(/* GraphQL */ `
