+++ packages/webapp/plop/crud/templates/components/itemList/listSkeleton/listSkeleton.component.hbs
@@ -1,5 +1,5 @@
 import { Card, CardContent } from '@sb/webapp-core/components/cards';
-import { Skeleton as SkeletonElement } from '@sb/webapp-core/components/skeleton';
+import { Skeleton as SkeletonElement } from '@sb/webapp-core/components/ui/skeleton';
 
 export const ListSkeleton = () => {
   return (
